Backport byte counting fix from 6.4.0 and 7.x.
The _chunk_actions function was counting characters instead of bytes.
This issue is fixed in ES-py 6.4.0 and 7.x, but we can't use the latter
due to cluster incompatibility and can't use the former due to
incompatibility with other libraries (such as dsl).

The 6.4.0 version uses elasticsearch6 as its package name and is
ignored by other libraries. Due to fear of other potentially breaking
changes, we've decided to fix it here, instad of changing the package
name for 6.4.0